# Conflicts:
#	deepfence_agent/build/plugincode-build.sh

+++ deepfence_agent/build/plugincode-build.sh
@@ -15,12 +15,12 @@ if [ $build_result -ne 0 ]
 then
     exit 1
 fi
-#make bin/SecretScanner
-#build_result=$?
-#if [ $build_result -ne 0 ]
-#then
-#    exit 1
-#fi
+make bin/SecretScanner
+build_result=$?
+if [ $build_result -ne 0 ]
+then
+    exit 1
+fi
 cd -
 rm -rf ./plugins/docker_bin
 mkdir ./plugins/docker_bin 2>/dev/null
